# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Live Score — Housing Trust Infrastructure API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/request-otp:
    post:
      operationId: requestOtp
      tags: [auth]
      summary: Request an OTP code for phone authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP code and receive auth token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerify"
      responses:
        "200":
          description: Auth token and user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout (client discards the bearer token)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "401":
          description: Authentication required
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OtpRequest:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
    OtpResponse:
      type: object
      required: [success, maskedPhone, expiresIn]
      properties:
        success:
          type: boolean
        maskedPhone:
          type: string
        expiresIn:
          type: integer
        devCode:
          type: string
    OtpVerify:
      type: object
      required: [phone, code]
      properties:
        phone:
          type: string
        code:
          type: string
    AuthToken:
      type: object
      required: [token, userId, name, phone, isAdmin]
      properties:
        token:
          type: string
        userId:
          type: integer
        name:
          type: string
        phone:
          type: string
        isAdmin:
          type: boolean
...