# lib/api-zod/src/generated/types/otpRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Live Score — Housing Trust Infrastructure API
 * OpenAPI spec version: 0.2.0
 */

export interface OtpRequest {
  phone: string;
}